Fix I_GetFreeMem for 64-bit platforms (backport of MR 1982)

+++ src/dummy/i_system.c
@@ -5,7 +5,7 @@ UINT8 graphics_started = 0;
 
 UINT8 keyboard_started = 0;
 
-UINT32 I_GetFreeMem(UINT32 *total)
+size_t I_GetFreeMem(size_t *total)
 {
 	*total = 0;
 	return 0;
@@ -179,4 +179,3 @@ char *I_ClipboardPaste(void)
 void I_RegisterSysCommands(void) {}
 
 #include "../sdl/dosstr.c"
-
